# internal/app/rest_api_v2/docs/swagger.yaml
basePath: /api/v2
definitions:
...
  HandlersHA.HaJail:
    properties:
      current_host:
        type: string
      jail_name:
        type: string
      latest_snapshot:
        type: string
      live:
        type: boolean
      parent_host:
        type: string
    type: object
  HandlersHA.HaVm:
    properties:
      current_host:
        type: string
      latest_snapshot:
        type: string
      live:
        type: boolean
      parent_host:
        type: string
      vm_name:
        type: string
    type: object
...
  RestApiConfig.HaNode:
    properties:
      address:
        type: string
      backup_node:
        type: boolean
      failover_strategy:
        type: string
      failover_time:
        type: integer
      hostname:
        type: string
      password:
        type: string
      port:
        type: string
      protocol:
        type: string
      registered:
        type: boolean
      startup_time:
        type: integer
      times_failed:
        type: integer
      user:
        type: string
    type: object
...
  handlers.SwaggerError:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  handlers.SwaggerSuccess:
    properties:
      message:
        description: success
        type: string
    type: object
...
info:
  contact: {}
  description: '`NOTE!` This REST API HTTP endpoint is located directly on the `Hoster`
    node.<br><br>The API should ideally be integrated into another system (e.g. a
    user-accessible back-end server), and not interacted with directly.<br><br>Please,
    take an extra care with the things you execute here, because some of them may
    be disruptive or non-revertible (e.g. vm destroy, snapshot rollback, host reboot,
    etc).'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Hoster Node REST API Docs
  version: "2.0"
paths:
  /ha/jail-list:
    get:
      description: Handle the HA enabled Jail list.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/HandlersHA.HaJail'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Handle the HA enabled Jail list.
      tags:
      - HA
...
  /ha/register:
    post:
      description: Handle the HA node registration.
      parameters:
      - description: Request payload
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/RestApiConfig.HaNode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SwaggerSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Handle the HA node registration.
      tags:
      - HA
  /ha/terminate:
    post:
      description: Handle the HA graceful termination signal.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SwaggerSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Handle the HA graceful termination signal.
      tags:
      - HA
  /ha/vm-list:
    get:
      description: Handle the HA enabled VM list.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/HandlersHA.HaVm'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Handle the HA enabled VM list.
      tags:
      - HA
...
swagger: "2.0"